requirements.txt
discord
psycopg2
pillow
numpy
requests
python-dotenv
matplotlib